Minor fixes for DQM classes.
A few months ago I did a simple test whether
having a huge library for all the DQM packages (instead of
.so per package) yield any performance results.

It did not.

However, I had to do some cleanup of the code in order to link it
correctly (mostly duplicate symbols):

1. L1TMenuHelper*.cc got de-duplicated;
2. some classes now have proper headers, instead of including .cc (yeah,
i know, crazy);
3. adding a few "static" keywords
