+++ DQM/L1TMonitor/BuildFile.xml
@@ -34,4 +34,8 @@
 <use   name="CondTools/L1Trigger"/>
 <use   name="root"/>
 <use   name="boost"/>
-<flags   EDM_PLUGIN="1"/>
+
+<export>
+  <flags   EDM_PLUGIN="1"/>
+  <lib name="DQML1TMonitor"/>
+</export>
